requirements.txt
# Core framework
hydra-core >= 1.3.2
torch
torchvision
torchaudio

# Computer vision and video processing
opencv-python >= 4.5.5.62
scipy >= 1.3.0
scikit-image >= 0.13.0
av >= 10.0.0
mediapipe
numpy >= 1.21.0

# Audio processing
pyaudio
openai-whisper

# Utilities
six >= 1.16.0
ollama
pydantic
keyboard

# Web interface (optional)
Flask >= 3.0.0
gTTS >= 2.5.1